Get available metrics with descriptions

Get available metrics with descriptions

get_available_metrics(pattern = NULL)

Arguments

pattern

Optional regex pattern to filter metrics

Value

Named character vector of metrics and descriptions

Examples

# Get all metrics
get_available_metrics()
#>                                                           views 
#>                            "Number of times videos were viewed" 
#>                                                        redViews 
#> "Number of times videos were viewed by YouTube Premium members" 
#>                                                    engagedViews 
#>            "Number of views that went past the initial seconds" 
#>                                                viewerPercentage 
#>        "Percentage of viewers who were logged in when watching" 
#>                                       videoThumbnailImpressions 
#>               "Number of times video thumbnails were displayed" 
#>                              videoThumbnailImpressionsClickRate 
#>                       "Click-through rate for video thumbnails" 
#>                                                           likes 
#>                                               "Number of likes" 
#>                                                        dislikes 
#>                                            "Number of dislikes" 
#>                                                        comments 
#>                                            "Number of comments" 
#>                                                          shares 
#>                            "Number of times videos were shared" 
#>                                               subscribersGained 
#>                                  "Number of subscribers gained" 
#>                                                 subscribersLost 
#>                                    "Number of subscribers lost" 
#>                                          videosAddedToPlaylists 
#>                "Number of times videos were added to playlists" 
#>                                      videosRemovedFromPlaylists 
#>            "Number of times videos were removed from playlists" 
#>                                         estimatedMinutesWatched 
#>                                         "Total minutes watched" 
#>                                      estimatedRedMinutesWatched 
#>                    "Minutes watched by YouTube Premium members" 
#>                                             averageViewDuration 
#>                    "Average duration of video views in seconds" 
#>                                           averageViewPercentage 
#>                           "Average percentage of video watched" 
#>                                                   playlistViews 
#>    "Number of times videos were viewed within playlist context" 
#>                                                  playlistStarts 
#>               "Number of times playlist playback was initiated" 
#>                                                   playlistSaves 
#>                      "Number of times users saved the playlist" 
#>                                           viewsPerPlaylistStart 
#>              "Average number of video views per playlist start" 
#>                                           averageTimeInPlaylist 
#>             "Average time in seconds viewers spent in playlist" 
#>                                     playlistAverageViewDuration 
#>                      "Average duration of playlist video views" 
#>                                 playlistEstimatedMinutesWatched 
#>                     "Total minutes watched in playlist context" 
#>                                                estimatedRevenue 
#>                                       "Estimated total revenue" 
#>                                              estimatedAdRevenue 
#>                                    "Estimated revenue from ads" 
#>                                      estimatedRedPartnerRevenue 
#>          "Estimated revenue from YouTube Premium subscriptions" 
#>                                              monetizedPlaybacks 
#>                                 "Number of monetized playbacks" 
#>                                                playbackBasedCpm 
#>                             "Cost per mille based on playbacks" 
#>                                                             cpm 
#>                           "Cost per mille based on impressions" 
#>                                                    grossRevenue 
#>                          "Gross revenue before revenue sharing" 
#>                                                   adImpressions 
#>                                      "Number of ad impressions" 
#>                                                 cardImpressions 
#>                                    "Number of card impressions" 
#>                                                      cardClicks 
#>                                         "Number of card clicks" 
#>                                                   cardClickRate 
#>                                  "Click-through rate for cards" 
#>                                           cardTeaserImpressions 
#>                             "Number of card teaser impressions" 
#>                                                cardTeaserClicks 
#>                                  "Number of card teaser clicks" 
#>                                             cardTeaserClickRate 
#>                           "Click-through rate for card teasers" 
#>                                           annotationImpressions 
#>                              "Number of annotation impressions" 
#>                                  annotationClickableImpressions 
#>                    "Number of clickable annotation impressions" 
#>                                                annotationClicks 
#>                                   "Number of annotation clicks" 
#>                                      annotationClickThroughRate 
#>                            "Click-through rate for annotations" 
#>                                   annotationClosableImpressions 
#>                     "Number of closable annotation impressions" 
#>                                                annotationCloses 
#>                                   "Number of annotation closes" 
#>                                             annotationCloseRate 
#>                         "Rate at which annotations were closed" 
#>                          membershipsCancellationSurveyResponses 
#>           "Number of completed membership cancellation surveys" 
#>                                        averageConcurrentViewers 
#>       "Average number of concurrent viewers during live stream" 
#>                                           peakConcurrentViewers 
#>          "Peak number of concurrent viewers during live stream" 
#>                                              audienceWatchRatio 
#>       "Ratio of views at each moment compared to initial views" 
#>                                    relativeRetentionPerformance 
#>              "Retention performance relative to similar videos" 
#>                                                 startedWatching 
#>                    "Number of views that started at this point" 
#>                                                 stoppedWatching 
#>                    "Number of views that stopped at this point" 
#>                                         totalSegmentImpressions 
#>                  "Total number of impressions for this segment" 

# Get only view-related metrics
get_available_metrics("view")
#>                                                           views 
#>                            "Number of times videos were viewed" 
#>                                                        redViews 
#> "Number of times videos were viewed by YouTube Premium members" 
#>                                                    engagedViews 
#>            "Number of views that went past the initial seconds" 
#>                                                viewerPercentage 
#>        "Percentage of viewers who were logged in when watching" 
#>                                             averageViewDuration 
#>                    "Average duration of video views in seconds" 
#>                                           averageViewPercentage 
#>                           "Average percentage of video watched" 
#>                                                   playlistViews 
#>    "Number of times videos were viewed within playlist context" 
#>                                           viewsPerPlaylistStart 
#>              "Average number of video views per playlist start" 
#>                                           averageTimeInPlaylist 
#>             "Average time in seconds viewers spent in playlist" 
#>                                     playlistAverageViewDuration 
#>                      "Average duration of playlist video views" 
#>                                        averageConcurrentViewers 
#>       "Average number of concurrent viewers during live stream" 
#>                                           peakConcurrentViewers 
#>          "Peak number of concurrent viewers during live stream" 
#>                                              audienceWatchRatio 
#>       "Ratio of views at each moment compared to initial views" 
#>                                                 startedWatching 
#>                    "Number of views that started at this point" 
#>                                                 stoppedWatching 
#>                    "Number of views that stopped at this point"