Get Reports
Source: R/get_report.R

Lists reporting jobs that have been scheduled for a channel or content owner.

get_report(ids, metrics, start_date = NULL, end_date = NULL,
  currency = NULL, dimensions, filters, historical_channel_data = NULL,
  max_results = NULL, sort = NULL, start_index = NULL, user_ip = NULL,
  ...)

Arguments

ids
Named vector with two potential names: channel or contentOwner If channel, potential values are: mine or channel_id If contentOwner, potential values are: owner_id of the content

metrics
String. Comma-separated list of YouTube Analytics metrics, such as views or likes,dislikes.

start_date
String. Must be in YYYY-MM-DD format.

end_date
String. Must be in YYYY-MM-DD format.

currency
Optional. String. Default is USD. Specifies what earnings metrics like earnings, adEarnings, grossRevenue, playbackBasedCpm, impressionBasedCpm will be reported in.

dimensions
String. Optional. Comma-separated list of YouTube Analytics dimensions, such as video or ageGroup,gender.

filters
Named Vector. Optional. For instance, ``video==pd1FJh59zxQ,Zhawgd0REhA;country==IT restricts the result set to include data for the given videos in Italy''

historical_channel_data
Boolean. Defaults is False. ``Whether the API response should include channels' watch time and view data from the time period prior to when the channels were linked to the content owner.''

max_results
Integer. Optional. The maximum number of rows to include in the response.

sort
String. Optional A comma-separated list of dimensions or metrics that determine the sort order for YouTube

start_index
Integer. Optional. ``The 1-based index of the first entity to retrieve.''

user_ip
``IP address of the end user for whom the API call is being made.''

…
Additional arguments passed to tubern_GET.

Value

named list

References

https://developers.google.com/youtube/analytics/v1/reference/reports/query

Examples

# NOT RUN {
get_report(ids = "channel==MINE", metrics = "views",
start_date = "2010-04-01", end_date ="2017-01-01")
# }